+++ NetworkSaver.h
@@ -9,7 +9,7 @@ public:
 	void SaveNetwork(NeuralNetwork* reseau, string filename);
 	NeuralNetwork* GetNetwork(string filename);
 	std::vector<Node> getNodes(size_t NbNodes);
-	std::vector<Connection> getConnections(int n_nodesOut);
+	std::vector<Connection> getConnections(size_t n_nodesOut);
 private:
 	CSVSaver CSV;
 };
